Convert settings (ANSI -> Unicode) on installation
The AHK ini functions does not handle UTF-8 encoding correctly. Either
ANSI or Unicode is recommended. The file was was therefore previously
stored as Unicode. Turns out Git diff/merge doesn't play well with that.

The file is now instead stored as ANSI to keep git functionality and
then converted to Unicode on install to support additional characters.